+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Changed
 - (#174) Update CI to build version based on tags
 - (#173) Use official Docker client
+- (#175) Changed the struct to use to start docker service
 
 #### Added
 - (#174) Add CHANGELOG.md file